backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy[asyncio]==2.0.23
alembic==1.12.1
psycopg[binary]==3.1.13
redis==5.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-decouple==3.8
pydantic-settings==2.0.3
prometheus-client==0.19.0
structlog==23.2.0
httpx==0.25.2